[prism] move test functions and tests to internal_test
